unsigned char in main.c
All variables and functions are, return, or reqest unsigned chars. Exceptions include:
The coins counter, needs to be big while able to be negative
The local j variable in updatemap(): it is an unsigned small because it needs to fill all of map.
A few other variables are voids
This is only in the main function